Yoav/config/rearrangement crate (#854)
