+++ crates/papyrus_config/Cargo.toml
@@ -12,4 +12,5 @@ thiserror.workspace = true
 [dev-dependencies]
 assert_matches.workspace = true
 itertools.workspace = true
+tempfile.workspace = true
 test_utils = { path = "../test_utils" }
