+++ crates/papyrus_config/resources/custom_config_example.json
