+++ crates/papyrus_config/src/command.rs
@@ -4,26 +4,24 @@ use std::path::PathBuf;
 use clap::{arg, value_parser, Arg, ArgMatches, Command};
 use serde_json::{json, Value};
 
-use crate::{update_config_map, SerializedParam, SubConfigError};
+use crate::loading::update_config_map;
+use crate::{ConfigError, ParamPath, SerializedParam};
 
-pub type ParamPath = String;
-pub type Description = String;
-
-pub fn get_command_matches(
+pub(crate) fn get_command_matches(
     config_map: &BTreeMap<ParamPath, SerializedParam>,
     command: Command,
     command_input: Vec<String>,
-) -> Result<ArgMatches, SubConfigError> {
+) -> Result<ArgMatches, ConfigError> {
     Ok(command.args(build_args_parser(config_map)).try_get_matches_from(command_input)?)
 }
 
-/// Takes matched arguments from the command line interface and env variables and updates the config
-/// map.
-/// Supports usize, bool and String.
-pub fn update_config_map_by_command_args(
+// Takes matched arguments from the command line interface and env variables and updates the config
+// map.
+// Supports usize, bool and String.
+pub(crate) fn update_config_map_by_command_args(
     config_map: &mut BTreeMap<ParamPath, SerializedParam>,
     arg_match: &ArgMatches,
-) -> Result<(), SubConfigError> {
+) -> Result<(), ConfigError> {
     for param_path_id in arg_match.ids() {
         let param_path = param_path_id.as_str();
         let new_value = get_arg_by_type(config_map, arg_match, param_path)?;
@@ -32,8 +30,8 @@ pub fn update_config_map_by_command_args(
     Ok(())
 }
 
-/// Builds the parser for the command line flags and env variables according to the types of the
-/// values in the config map.
+// Builds the parser for the command line flags and env variables according to the types of the
+// values in the config map.
 fn build_args_parser(config_map: &BTreeMap<ParamPath, SerializedParam>) -> Vec<Arg> {
     let mut args_parser = vec![
         // Custom_config_file_path.
@@ -60,12 +58,12 @@ fn build_args_parser(config_map: &BTreeMap<ParamPath, SerializedParam>) -> Vec<A
     args_parser
 }
 
-/// Converts clap arg_matches into json values.
+// Converts clap arg_matches into json values.
 fn get_arg_by_type(
     config_map: &BTreeMap<ParamPath, SerializedParam>,
     arg_match: &ArgMatches,
     param_path: &str,
-) -> Result<Value, SubConfigError> {
+) -> Result<Value, ConfigError> {
     match config_map[param_path].value {
         Value::Number(_) => Ok(json!(arg_match.try_get_one::<usize>(param_path)?)),
         Value::Bool(_) => Ok(json!(arg_match.try_get_one::<bool>(param_path)?)),
