+++ crates/papyrus_config/src/config_test.rs
@@ -10,16 +10,18 @@ use serde_json::json;
 use test_utils::get_absolute_path;
 
 use crate::command::{get_command_matches, update_config_map_by_command_args};
-use crate::{
-    append_sub_config_name, combine_config_map_and_pointers, deserialize_milliseconds_to_duration,
-    get_maps_from_raw_json, load, ser_param, update_config_map_by_custom_config,
-    update_config_map_by_pointers, ParamPath, PointerParam, SerializeConfig, SerializedParam,
-    SubConfigError,
+use crate::converters::deserialize_milliseconds_to_duration;
+use crate::dumping::{
+    append_sub_config_name, combine_config_map_and_pointers, ser_param, SerializeConfig,
 };
+use crate::loading::{
+    get_maps_from_raw_json, load, update_config_map_by_custom_config, update_config_map_by_pointers,
+};
+use crate::{ConfigError, ParamPath, PointerParam, SerializedParam};
 
 #[derive(Clone, Serialize, Deserialize, Debug, PartialEq)]
-pub struct InnerConfig {
-    pub a: usize,
+struct InnerConfig {
+    a: usize,
 }
 
 impl SerializeConfig for InnerConfig {
@@ -29,8 +31,8 @@ impl SerializeConfig for InnerConfig {
 }
 
 #[derive(Clone, Serialize, Deserialize, Debug, PartialEq)]
-pub struct OptionalConfig {
-    pub o: usize,
+struct OptionalConfig {
+    o: usize,
 }
 
 impl SerializeConfig for OptionalConfig {
@@ -40,11 +42,11 @@ impl SerializeConfig for OptionalConfig {
 }
 
 #[derive(Clone, Serialize, Deserialize, Debug, PartialEq)]
-pub struct OuterConfig {
-    pub inner: InnerConfig,
-    pub b: usize,
-    pub some_optional: Option<OptionalConfig>,
-    pub none_optional: Option<OptionalConfig>,
+struct OuterConfig {
+    inner: InnerConfig,
+    b: usize,
+    some_optional: Option<OptionalConfig>,
+    none_optional: Option<OptionalConfig>,
 }
 
 impl SerializeConfig for OuterConfig {
@@ -96,11 +98,11 @@ fn dump_and_load_config() {
 }
 
 #[derive(Clone, Serialize, Deserialize, Debug, PartialEq)]
-pub struct TypicalConfig {
+struct TypicalConfig {
     #[serde(deserialize_with = "deserialize_milliseconds_to_duration")]
-    pub a: Duration,
-    pub b: String,
-    pub c: bool,
+    a: Duration,
+    b: String,
+    c: bool,
 }
 
 impl SerializeConfig for TypicalConfig {
@@ -140,11 +142,10 @@ fn test_pointers_flow() {
         ser_param("a2", &json!(5), "This is a."),
     ]);
     let pointers = vec![(
-        "common_a".to_owned(),
-        "This is common a".to_owned(),
+        ser_param("common_a", &json!(10), "This is common a"),
         vec!["a1".to_owned(), "a2".to_owned()],
     )];
-    let stored_map = combine_config_map_and_pointers(config_map, pointers).unwrap();
+    let stored_map = combine_config_map_and_pointers(config_map, &pointers).unwrap();
     assert_eq!(
         stored_map["a1"],
         json!(PointerParam {
@@ -155,14 +156,14 @@ fn test_pointers_flow() {
     assert_eq!(stored_map["a2"], stored_map["a1"]);
     assert_eq!(
         stored_map["common_a"],
-        json!(SerializedParam { description: "This is common a".to_owned(), value: json!(5) })
+        json!(SerializedParam { description: "This is common a".to_owned(), value: json!(10) })
     );
 
     let serialized = serde_json::to_string(&stored_map).unwrap();
     let loaded = serde_json::from_str(&serialized).unwrap();
     let (mut loaded_config_map, loaded_pointers_map) = get_maps_from_raw_json(loaded);
     update_config_map_by_pointers(&mut loaded_config_map, &loaded_pointers_map).unwrap();
-    assert_eq!(loaded_config_map["a1"].value, json!(5));
+    assert_eq!(loaded_config_map["a1"].value, json!(10));
     assert_eq!(loaded_config_map["a1"], loaded_config_map["a2"]);
 }
 
@@ -176,14 +177,15 @@ fn test_replace_pointers() {
     assert_eq!(config_map["a"], config_map["c"]);
 
     let err = update_config_map_by_pointers(&mut BTreeMap::default(), &pointers_map).unwrap_err();
-    assert_matches!(err, SubConfigError::PointerTargetNotFound { .. });
+    assert_matches!(err, ConfigError::PointerTargetNotFound { .. });
 }
 
 #[test]
-pub fn test_update_by_custom_config() {
+fn test_update_by_custom_config() {
     let mut config_map =
         BTreeMap::from([ser_param("param_path", &json!("default value"), "This is a.")]);
-    let custom_config_path = get_absolute_path("config/custom_config_example.json");
+    let custom_config_path =
+        get_absolute_path("crates/papyrus_config/resources/custom_config_example.json");
     update_config_map_by_custom_config(&mut config_map, &custom_config_path).unwrap();
     assert_eq!(config_map["param_path"].value, json!("custom value"));
 }
