+++ crates/papyrus_node/src/main.rs
@@ -44,7 +44,7 @@ async fn run_threads(config: Config) -> anyhow::Result<()> {
         storage_reader: StorageReader,
         storage_writer: StorageWriter,
     ) -> Result<(), StateSyncError> {
-        let Some(sync_config) = config.sync else {return Ok(())};
+        let Some(sync_config) = config.sync else { return Ok(()) };
         let central_source =
             CentralSource::new(config.central.clone(), VERSION_FULL, storage_reader.clone())
                 .map_err(CentralError::ClientCreation)?;
