+++ crates/papyrus_storage/src/db/mod.rs
@@ -10,7 +10,8 @@ use std::result;
 use std::sync::Arc;
 
 use libmdbx::{Cursor, DatabaseFlags, Geometry, WriteFlags, WriteMap};
-use papyrus_config::{ser_param, ParamPath, SerializeConfig, SerializedParam, DEFAULT_CHAIN_ID};
+use papyrus_config::dumping::{ser_param, SerializeConfig};
+use papyrus_config::{ParamPath, SerializedParam};
 use serde::{Deserialize, Serialize};
 use starknet_api::core::ChainId;
 
@@ -45,7 +46,7 @@ impl Default for DbConfig {
     fn default() -> Self {
         DbConfig {
             path_prefix: PathBuf::from("./data"),
-            chain_id: ChainId(DEFAULT_CHAIN_ID.to_string()),
+            chain_id: ChainId("SN_MAIN".to_string()),
             min_size: 1 << 20,    // 1MB
             max_size: 1 << 40,    // 1TB
             growth_step: 1 << 26, // 64MB
@@ -239,7 +240,7 @@ impl<'env, 'txn, K: StorageSerde, V: StorageSerde> TableHandle<'env, K, V> {
         // TODO: Support zero-copy. This might require a return type of Cow<'env, ValueType>.
         let bin_key = key.serialize()?;
         let Some(bytes) = txn.txn.get::<Cow<'env, [u8]>>(&self.database, &bin_key)? else {
-            return Ok(None)
+            return Ok(None);
         };
         let value = V::deserialize(&mut bytes.as_ref()).ok_or(DbError::InnerDeserialization)?;
         Ok(Some(value))
